Relocate SmoothOperator to standard provider and add tests
