+++ providers/standard/provider.yaml
@@ -48,7 +48,7 @@ operators:
       - airflow.providers.standard.operators.empty
       - airflow.providers.standard.operators.trigger_dagrun
       - airflow.providers.standard.operators.latest_only
-
+      - airflow.providers.standard.operators.smooth
 sensors:
   - integration-name: Standard
     python-modules:
